# backend-rust/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ABP Studio Rust API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/{user_id}": {
      "put": {
        "operationId": "update_user",
        "summary": "Migrated from Python update_user",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "operationId": "delete_user",
        "summary": "Migrated from Python delete_user",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/gallery/images/batch-delete": {
      "post": {
        "operationId": "batch_delete_images",
        "summary": "Migrated from Python batch_delete_images",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/gallery/videos/batch-delete": {
      "post": {
        "operationId": "batch_delete_videos",
        "summary": "Migrated from Python batch_delete_videos",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/gallery/images/batch-share": {
      "post": {
        "operationId": "batch_share_images",
        "summary": "Migrated from Python batch_share_images",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/gallery/videos/batch-share": {
      "post": {
        "operationId": "batch_share_videos",
        "summary": "Migrated from Python batch_share_videos",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/gallery/images/share-all": {
      "post": {
        "operationId": "share_all_images",
        "summary": "Migrated from Python share_all_images",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/gallery/videos/share-all": {
      "post": {
        "operationId": "share_all_videos",
        "summary": "Migrated from Python share_all_videos",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/gallery/images/batch-download": {
      "post": {
        "operationId": "batch_download_images",
        "summary": "Migrated from Python batch_download_images",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/gallery/videos/batch-download": {
      "post": {
        "operationId": "batch_download_videos",
        "summary": "Migrated from Python batch_download_videos",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}